+++ src/components/common/index.ts
@@ -1,3 +1,2 @@
 export * from './BottomSheet';
 export * from './Navbar';
-export * from './Template';
