+++ src/components/index.ts
@@ -1,5 +1,2 @@
 export * from './auth';
-export * from './businesscard';
 export * from './common';
-export * from './maintabview';
-export * from './petcard';
