+++ src/components/templates/MainTabView/BusinessCardView.tsx
