+++ src/components/templates/ProgressView/GenerateItem.tsx
