+++ src/components/templates/ProgressView/GenerateView.tsx
