+++ src/components/templates/ProgressView/ProgressBar.tsx
