+++ src/components/templates/ProgressView/index.ts
@@ -1,4 +1,3 @@
 export * from './GenerateItem';
 export * from './GenerateView';
 export * from './ProgressBar';
-export * from './Tab';
