+++ src/components/templates/index.ts
@@ -1,0 +1,3 @@
+export * from './MainTabView';
+export * from './ProgressView';
+export * from './Template';
