+++ src/types/svg.d.ts
@@ -1,6 +1,4 @@
 declare module '*.svg' {
   import React = require('react');
   export const ReactComponent: React.FC<React.SVGProps<SVGSVGElement>>;
-  const src: string;
-  export default src;
 }
